+++ R/mod_table.R
@@ -44,13 +44,11 @@ mod_china_server <- function(input, output, session, df){
       dplyr::arrange(-confirmed) %>% 
       dplyr::mutate(
         confirmed = as.integer(confirmed),
-        recovered = as.integer(recovered),
         death = as.integer(death)
       ) %>% 
       dplyr::select(
         Province = state,
         Confirmed = confirmed, 
-        Recovered = recovered,
         Deaths = death
       ) %>% 
       as_f7_table(card = TRUE)