+++ docs/news.md
@@ -43,8 +43,8 @@
 ## coronavirus 0.0.3
 
 - Connect all charts in dxy and weixin tabs
-- Corrected deaths and recovered in DingXiangYing data, was swapped, see [#2](https://github.com/JohnCoene/coronavirus/issues/2)
-- Number of suspected by city given by DingXiangYing is wildly inaccurate, has been removed.
+- Corrected deaths and recovered in DingXiangYuan data, was swapped, see [#2](https://github.com/JohnCoene/coronavirus/issues/2)
+- Number of suspected by city given by DingXiangYuan is wildly inaccurate, has been removed.
 - JHU timeline uses time scale for better, more accurate representation of events.
 - Corrected legend on timeline in JHU tab